Added timeout when waiting for the controller reply.

+++ lib/serial/actions.rb
@@ -12,18 +12,23 @@ module Serial
       
         ################################################################################
         def state
-          task = Task.create! :house => self.house,
-                          :operation => {
+          task = Task.create! :house  => self.house,
+                          :operation  => {
                             :sent => {
-                              :device_identification => self.device.identification,
-                              :action_command => self.command,
-                              :action_query_state => self.query_state
+                              :device_identification  => self.device.identification,
+                              :action_command         => self.command,
+                              :action_query_state     => self.query_state
                             }
                           }
 
-          task.wait_for_response
+          begin
+            task.wait_for_response
+          rescue Serial::Error::UnknownSerialError
+            response = {:code           => range? ? 0 : 'state_off',
+                        :error_message  => I18n.t(:error_obtaining_the_action_status, :scope => [:action, :messages, :error])}
+          end
 
-          response = {:code => task.answered_status}
+          response = {:code => task.answered_status} unless response.is_a?(Hash)
 
           if range?
             response[:state] = response[:code]
@@ -31,7 +36,7 @@ module Serial
             response[:state] = I18n.t(response[:code].to_sym, :scope => [:action, :device, :turn_on_off, :state])
           end
 
-          task.destroy
+          task.destroy unless response.key?(:error_message)
           response
         end
 
@@ -40,16 +45,20 @@ module Serial
           task = Task.create! :house => self.house,
                     :operation => {
                       :sent => {
-                        :device_identification => self.device.identification,
-                        :action_command => self.command,
-                        :value => self.new_value
+                        :device_identification  => self.device.identification,
+                        :action_command         => self.command,
+                        :value                  => self.new_value
                       }
                     }
 
-          task.wait_for_response
+          begin
+            task.wait_for_response
+          rescue Serial::Error::UnknownSerialError
+            response = {:success => 500, :error => true}
+          end
 
-          response = {:success => task.answered_setting_value.to_i, :operation => task.operation}
-          task.destroy
+          response = {:success => task.answered_setting_value.to_i, :operation => task.operation} unless response.is_a?(Hash)
+          task.destroy unless response.key?(:error)
           
           response
         end
